Deduplicate Quick Lane source across dev and dist
